JBIDE-19080: Rename org.jboss.tools.hibernate.spi into org.jboss.tools.hibernate.runtime.spi
