+++ plugins/org.jboss.tools.birt.oda.ui/META-INF/MANIFEST.MF
@@ -16,7 +16,7 @@ Require-Bundle: org.eclipse.ui,
  org.hibernate.eclipse.console,
  org.hibernate.eclipse;bundle-version="3.8.0",
  org.eclipse.core.resources,
- org.jboss.tools.hibernate.spi;bundle-version="4.0.0";resolution:=optional,
+ org.jboss.tools.hibernate.runtime.spi;bundle-version="4.0.0";resolution:=optional,
  org.jboss.tools.hibernate3_5;bundle-version="4.0.0";resolution:=optional
 Bundle-RequiredExecutionEnvironment: J2SE-1.5
 Bundle-ActivationPolicy: lazy
