+++ plugins/org.jboss.tools.birt.oda.ui/src/org/jboss/tools/birt/oda/ui/impl/HibernateSelectionPageHelper.java
@@ -37,7 +37,7 @@ import org.hibernate.console.KnownConfigurations;
 import org.jboss.tools.birt.oda.IOdaFactory;
 import org.jboss.tools.birt.oda.ui.Activator;
 import org.jboss.tools.birt.oda.ui.Messages;
-import org.jboss.tools.hibernate.spi.ISessionFactory;
+import org.jboss.tools.hibernate.runtime.spi.ISessionFactory;
 
 /**
  * Helper class for Hibernate wizard and property page
