+++ plugins/org.jboss.tools.birt.oda/META-INF/MANIFEST.MF
@@ -12,7 +12,7 @@ Require-Bundle: org.eclipse.core.runtime,
  org.eclipse.datatools.connectivity.oda;bundle-version="[3.1.0,4.0.0)",
  org.eclipse.datatools.connectivity.oda.profile;bundle-version="[3.0.4,4.0.0)",
  org.hibernate.eclipse;bundle-version="3.8.0";resolution:=optional,
- org.jboss.tools.hibernate.spi;bundle-version="4.0.0";resolution:=optional,
+ org.jboss.tools.hibernate.runtime.spi;bundle-version="4.0.0";resolution:=optional,
  org.jboss.tools.hibernate3_5;bundle-version="4.0.0";resolution:=optional
 Bundle-ActivationPolicy: lazy
 Bundle-RequiredExecutionEnvironment: J2SE-1.5
