+++ plugins/org.jboss.tools.birt.oda/src/org/jboss/tools/birt/oda/Activator.java
@@ -15,7 +15,7 @@ import org.eclipse.core.runtime.IStatus;
 import org.eclipse.core.runtime.Plugin;
 import org.eclipse.core.runtime.Status;
 import org.hibernate.SessionFactory;
-import org.jboss.tools.hibernate.spi.ISessionFactory;
+import org.jboss.tools.hibernate.runtime.spi.ISessionFactory;
 import org.osgi.framework.BundleContext;
 
 /**
